Configure and Processing tabs. 
Progress indicator: show current file, hex info, step while processing
Interprocess communication with Queue

+++ gui/app.py
@@ -1,46 +1,14 @@
-import multiprocessing
+from multiprocessing import Process, Queue, JoinableQueue
+import queue
 
-from tkinter import filedialog, Label
+from tkinter import filedialog, Label, messagebox
 import customtkinter
 
 from config import CONFIG
 from manager import start_manager
 
-# FIXME not safe to change config while process is running.
-#   not sure how Python synchronizes changes to module objects like CONFIG betweens processes.
-
-def select_raw_directory(raw_path_label):
-    """Get the raw directory with button click (default assigned to local directory)"""
-    print("Raw Directory Button clicked!")
-    raw_directory_selected = filedialog.askdirectory()
-    CONFIG["RAW_PATH"] = raw_directory_selected
-    raw_path_label.configure(text=CONFIG["RAW_PATH"])
-
-
-def select_processing_directory(PROCESSING_PATH_label):
-    """Get the processing directory with button click (default assigned to local directory)"""
-    print("Processing Directory Button clicked!")
-    processing_directory_selected = filedialog.askdirectory()
-    print(processing_directory_selected)
-    CONFIG["PROCESSING_PATH"] = (processing_directory_selected)
-    PROCESSING_PATH_label.configure(text=CONFIG["PROCESSING_PATH"])
-
-
-def select_config_directory(config_path_label):
-    """Get the config directory with button click (default assigned to local directory)"""
-    print("Configuration Directory Button clicked!")
-    config_directory_selected = filedialog.askdirectory()
-    CONFIG["CTD_CONFIG_PATH"] = config_directory_selected
-    config_path_label.configure(text=CONFIG["CTD_CONFIG_PATH"])
-
-
-def select_database_directory(database_path_label):
-    """Get the database directory with button click (default assigned to local directory)"""
-    print("Database Directory Button clicked!")
-    database_directory_selected = filedialog.askdirectory()
-    database_path = database_directory_selected
-    CONFIG["CTD_DATABASE_PATH"] = database_directory_selected
-    database_path_label.configure(text=CONFIG["CTD_DATABASE_PATH"])
+from .processing_panel import ProcessingPanel
+from .config_panel import ConfigPanel
 
 class App:
     def __init__(self) -> None:
@@ -51,19 +19,38 @@ class App:
         if self.proc is not None and self.proc.is_alive():
             raise Exception("existing process is running")
 
-        self.proc = multiprocessing.Process(target=start_manager, args=())
+        # for messages sent by worker processes.
+        self.recv = recv = Queue()
+
+        # our recv is the other process' send
+        self.proc = Process(target=start_manager, args=(recv,))
         self.proc.start()
 
+        self.processing_panel.set_processing_state(True)
+
+        self._after_id = self.window.after_idle(self.process_events)
 
     # Terminate the process
-    def stop_process(self):
+    def stop_process(self, join_timeout=None):
         """Stop processing with a button click"""
+
+        self.window.after_cancel(self._after_id)
+
         if self.proc is None:
             print("No processing started.")
             return
 
-        self.proc.terminate()  # sends a SIGTERM
+        if join_timeout is not None:
+            self.proc.join(join_timeout)
+
+        if self.proc.is_alive():
+            print("Terminating process")
+            self.proc.terminate()  # sends a SIGTERM
+
         self.proc = None
+
+        self.processing_panel.set_processing_state(False)
+
         print("Stopped processing.")
         print("Temporary files may remain in the raw directory due to cancelled processing.")
 
@@ -71,6 +58,50 @@ class App:
         #  print(file_name)
         # print(base_file_name)
         #thought process here to check if these two are equal and if not, delete file_name file
+
+    def process_events(self):
+        """Process events from subprocess and monitor if it's alive"""
+        try:
+            while True:
+                msg = self.recv.get(block=False)
+                self.process_msg(msg)
+
+        except queue.Empty:
+            pass
+
+        if self.proc is not None:
+            if self.proc.is_alive():
+                # schedule next process_events in 100ms
+                self._after_id = self.window.after(100, self.process_events)
+            else:
+                # cleanup process state
+                self.stop_process()
+
+    def process_msg(self, msg):
+        print("msg:", msg)
+        label = msg[0]
+
+        if label == "process_step":
+            self.processing_panel.set_step(msg[1], msg[2], msg[3])
+        elif label == "begin":
+            self.processing_panel.reset_progress(msg[1])
+        elif label == "start":
+            _, name, i, num_pending = msg
+            self.processing_panel.start_file(name, i, num_pending)
+        elif label == "hex_info":
+            self.processing_panel.set_file_info(msg[1], msg[2])
+        elif label == "finish":
+            _, name, i, num_processed = msg
+            self.processing_panel.finished_file(name, i, num_processed)
+        elif label == "done":
+            self.stop_process(2_000)
+        elif label == "usermsg":
+            messagebox.showinfo("CTD Processing Message", msg[1])
+        elif label == "error":
+            messagebox.showerror("CTD Processing Error", msg[1])
+        else:
+            print("WARNING: unknown message", msg)
+
 
     def build(self):
         PROCESSING_PATH = CONFIG["PROCESSING_PATH"]
@@ -87,54 +118,19 @@ class App:
 
         window.title("Seabird CTD Processor")
 
-        self.build_config(window)
+        self.tabview = tabview = customtkinter.CTkTabview(window, command=self.on_tab_change)
+        tabview.pack(fill="both", expand=True)
+        # CTkFrame
+        config_tab = tabview.add("Configure")
+        processing_tab = tabview.add("Processing")
 
-        # process stage 1 button
-        process_button = customtkinter.CTkButton(
-            window, text="Process Data", font=("Arial", 30, 'bold'), fg_color="#5D892B", hover_color="#334B18",
-            command=lambda: self.start_process()
-        ).pack(pady=(10,10))
-        stage1_path_label = customtkinter.CTkLabel(window, text="Process data from .hex to BinDown stage", font=CONFIG["LABEL_FONTS"])
-        stage1_path_label.pack(pady=(5, 25))
+        self.config_panel = ConfigPanel(config_tab)
+        self.processing_panel = ProcessingPanel(processing_tab, self)
 
-        # Stop process button
-        stop_button = customtkinter.CTkButton(
-            window, text="Stop", font=("Arial", 20, 'bold'), fg_color="#AC3535", hover_color="#621E1E",
-            command=lambda: self.stop_process()
-        ).pack(pady=(0,20))
-
-    def build_config(self, window):
-        # raw directory button
-        raw_directory_button = customtkinter.CTkButton(
-            window, text="Select Raw Directory", font=CONFIG["LABEL_FONTS"],
-            command=lambda: select_raw_directory(raw_path_label)
-        ).pack()
-        raw_path_label = customtkinter.CTkLabel(window, text=CONFIG["RAW_PATH"], font=CONFIG["LABEL_FONTS"])
-        raw_path_label.pack(pady=(5, 25))
-
-        # processing directory button
-        processing_directory_button = customtkinter.CTkButton(
-            window, text="Select processing Directory", font=CONFIG["LABEL_FONTS"],
-            command=lambda: select_processing_directory(PROCESSING_PATH_label)
-        ).pack()
-        PROCESSING_PATH_label = customtkinter.CTkLabel(window, text=CONFIG["PROCESSING_PATH"], font=CONFIG["LABEL_FONTS"])
-        PROCESSING_PATH_label.pack(pady=(5, 25))
-
-        # configuration directory button
-        config_directory_button = customtkinter.CTkButton(
-            window, text="Select Configuration Directory", font=CONFIG["LABEL_FONTS"],
-            command=lambda: select_config_directory(config_path_label)
-        ).pack()
-        config_path_label = customtkinter.CTkLabel(window, text=CONFIG["CTD_CONFIG_PATH"], font=CONFIG["LABEL_FONTS"])
-        config_path_label.pack(pady=(5, 25))
-
-        # database directory button
-        database_directory_button = customtkinter.CTkButton(
-            window, text="Select Database Directory", font=CONFIG["LABEL_FONTS"],
-            command=lambda: select_database_directory(database_path_label)
-        ).pack()
-        database_path_label = customtkinter.CTkLabel(window, text=CONFIG["DATABASE_MDB_FILE"], font=CONFIG["LABEL_FONTS"])
-        database_path_label.pack(pady=(5, 25))
+    def on_tab_change(self):
+        current = self.tabview.get()
+        if current == "Processing":
+            self.processing_panel.prepare()
 
     def start(self):
         "Start the tkinter event loop"
